adding catch if redis doesnt exist for driver
