+++ gamechangerml/api/fastapi/routers/controls.py
@@ -187,7 +187,7 @@ async def download_corpus(corpus_dict: dict, response: Response):
     return get_process_status()
 
 @router.post("/trainModel", status_code=200)
-async def tain_model(model_dict: dict, response: Response):
+async def train_model(model_dict: dict, response: Response):
     """load_latest_models - endpoint for updating the transformer model
     Args:
         model_dict: dict; {"encoder_model":"msmarco-distilbert-base-v2", "gpu":true, "upload":false,"version": "v5"}
