+++ gamechangerml/api/utils/processmanager.py
@@ -7,41 +7,48 @@ corpus_download = "corpus_download"
 training_progress = "training_progress"
 
 # the dictionary that holds all the progress values
-PROCESS_STATUS = CacheVariable("process_status", True)
-COMPLETED_PROCESS = CacheVariable("completed_process", True)
-thread_lock = threading.Lock()
+try:
+    PROCESS_STATUS = CacheVariable("process_status", True)
+    COMPLETED_PROCESS = CacheVariable("completed_process", True)
+    thread_lock = threading.Lock()
 
-PROCESS_STATUS.value = {}
-COMPLETED_PROCESS.value = []
+    PROCESS_STATUS.value = {}
+    COMPLETED_PROCESS.value = []
+except Exception as e:
+    print("Could not get process manager")
 
 def update_status(key, progress, total):
     # log update at most 1000 times
-    update_step = total/1000
-    if update_step < 1:
-        update_step = 1
-        
-    if progress == total:
-        date = datetime.now()
-        date_string = date.strftime("%Y-%m-%d %H:%M:%S")
-        completed = {
-            "process": key,
-            "total": total,
-            "date": date_string
-        }
-        with thread_lock:
-            if key in PROCESS_STATUS.value:
-                temp = PROCESS_STATUS.value
-                temp.pop(key, None)
-                PROCESS_STATUS.value = temp
-            completed_list = COMPLETED_PROCESS.value
-            completed_list.append(completed)
-            COMPLETED_PROCESS.value = completed_list
-    elif progress%update_step ==0:
-        status = {
-            "progress":progress,
-            "total":total
-        }
-        with thread_lock:
-            status_dict = PROCESS_STATUS.value
-            status_dict[key] = status
-            PROCESS_STATUS.value = status_dict
+    try:
+        update_step = total/1000
+        if update_step < 1:
+            update_step = 1
+            
+        if progress == total:
+            date = datetime.now()
+            date_string = date.strftime("%Y-%m-%d %H:%M:%S")
+            completed = {
+                "process": key,
+                "total": total,
+                "date": date_string
+            }
+            with thread_lock:
+                if key in PROCESS_STATUS.value:
+                    temp = PROCESS_STATUS.value
+                    temp.pop(key, None)
+                    PROCESS_STATUS.value = temp
+                completed_list = COMPLETED_PROCESS.value
+                completed_list.append(completed)
+                COMPLETED_PROCESS.value = completed_list
+        elif progress%update_step ==0:
+            status = {
+                "progress":progress,
+                "total":total
+            }
+            with thread_lock:
+                status_dict = PROCESS_STATUS.value
+                status_dict[key] = status
+                PROCESS_STATUS.value = status_dict
+    except Exception as e:
+            print(e)
+            print("Could not update status")
